resource and struct type decoding encoding and tests
